# api.yaml
swagger: "2.0"
info:
  title: This Swagger UI Docs
  description: This is my company documentation made by me
  version: 1.0
  contact:
    email: iitjeeabhishek@gmail.com
  license:
    name: Apache 2.0
    url: "http://apache.org/"
  servers:
    -url: "http://localhost:4000/"
    description: This is my local server
    -url: "http://company.com/"
    description: This is my company server

schemes: [http, https]

paths:
# Returns a string
  /string: 
    get:
      summary: returns a string
      description: It is just returning a fake string
      responses:
        200:
          description: Success
          schema:
            type: string
            example: "A fake string"
        400:
          description: Not Found
          schema:
            type: string
            example: "A fake string"
        500:
          description: Internal Server Error
          schema:
            type: string
            example: "A fake string"

#returns a object
  /user: 
    get:
      summary: returns a user
      description: It is just returning a fake user
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        400:
          description: Not Found
          schema:
            type: object
        500:
          description: Internal Server Error
          schema:
            type: object

# Returning an arrray of users
  /users: 
    get:
      summary: returns array of users
      description: Array returns
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
        400:
          description: Not Found
          schema:
            type: object
        500:
          description: Internal Server Error
          schema:
            type: object
  
  /users/{userId}:
    get:
      summary: returns array of users
      description: Array returns
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
  
  /create:
    post:
      summary: this is for the posting data
      description: this is a description
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
      - in: body
        name: user
        schema:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
      responses:
        200:
          description: Ok
        400: 
          description: Not found